+++ src/components/cards/Cards.css
@@ -11,85 +11,6 @@
   gap: 20px;
 }
 
-.Cards .card {
-  display: flex;
-  flex-direction: column;
-  border-radius: 15px;
-  overflow: hidden;
-  box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
-  text-decoration: none;
-}
-
-.Cards .card_image_container {
-  height: 200px;
-  display: flex;
-  align-items: center;
-  overflow: hidden;
-}
-
-.Cards .card_image {
-  height: 100%;
-  width: 100%;
-  object-fit: cover;
-  transition: all 0.2s linear;
-  transform: scale(1);
-}
-
-.Cards .card_image:hover {
-  transition: all 0.2s linear;
-  transform: scale(1.3);
-}
-
-.Cards .card_information {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  height: 135px;
-  padding: 20px;
-}
-
-.Cards .card_menu_container {
-  display: flex;
-  justify-content: space-between;
-}
-
-.Cards .createdAt {
-  color: #666666;
-  font-family: Pretendard;
-  font-size: 13px;
-  font-weight: 400;
-}
-
-.Cards .kebab_container {
-  position: relative;
-}
-
-.Cards .popover_container {
-  position: absolute;
-}
-
-.Cards .description {
-  font-family: Pretendard;
-  font-size: 16px;
-  font-weight: 400;
-  color: black;
-  line-height: 24px;
-  display: -webkit-box;
-  -webkit-line-clamp: 2;
-  -webkit-box-orient: vertical;
-  overflow: hidden;
-}
-
-.Cards .no_link_msg {
-  grid-column: 2;
-  text-align: center;
-  margin: 41px auto 35px;
-  font-family: Pretendard;
-  font-size: 16px;
-  font-weight: 400;
-  line-height: 24px;
-}
-
 /* tablet */
 @media (max-width: 1199px) {
   .Cards .container {
